LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/forward-beforeunload-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back


Harness Error (TIMEOUT), message = null

TIMEOUT forward() inside onbeforeunload Test timed out
